requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
aiohttp==3.9.1